Consolidated files in the "Scripts" Directory
I've moved all the files into the same scripts directory so that there
no longer exists the replication of python files.

+++ Scripts/addLinks.py
@@ -8,12 +8,12 @@ This tool only works with shapefiles or zip files whose   '''
 import arcpy, os, zipfile
 
 dir = arcpy.GetParameterAsText(0)
-fc = arcpy.GetParameterAsText(1) //optional
+fc = arcpy.GetParameterAsText(1) #optional
 url = arcpy.GetParameterAsText(2)
 
-//NEW MOD (V1.3) - check if input class if a zipfile.  If it is, 
-zipBool = arcpy.GetParameter(3) //optional
-zipFile = arcpy.GetParameterAsText(4) //optional
+#NEW MOD (V1.3) - check if input class if a zipfile.  If it is, 
+zipBool = arcpy.GetParameter(3) #optional
+zipFile = arcpy.GetParameterAsText(4) #optional
 if zipBool:
     fc = zipFile
     fc_folder = os.path.dirname(os.path.realpath(fc))
@@ -22,30 +22,27 @@ if zipBool:
         z.extractall(newPath)
     os.remove(fc)
     
-    //Add the fields and populate them
+    #Add the fields and populate them
     list = os.listdir(newPath)
     if list[0][:-3] == list[1][:-3]:
     
         fc = os.listdir(newPath)[0][:-3] + "shp"
         cursor = arcpy.da.UpdateCursor(fc, ("filename","location1","location2"))
-    //else:
-        //give an error message
+        for row in cursor:
+            file = row[0]
+            if os.path.exists(dir + "\\" + row[0] + "Crop.zip"):
+                row[1] = url + "/" + row[0] + "Crop.zip"
+            if os.path.exists(dir + "\\" + row[0] + "White.zip"):
+                row[2] = url + "/" + row[0] + "White.zip"
+            cursor.updateRow(row)
 
-for row in cursor:
-    file = row[0]
-    if os.path.exists(dir + "\\" + row[0] + "Crop.zip"):
-        row[1] = url + "/" + row[0] + "Crop.zip"
-    if os.path.exists(dir + "\\" + row[0] + "White.zip"):
-        row[2] = url + "/" + row[0] + "White.zip"
-    cursor.updateRow(row)
-
-//Zip the results.  
-
-list = os.listdir(newPath)         
+#Zip the results.  
+if zipBool:
+    list = os.listdir(newPath)         
     with zipfile.ZipFile(my_zip, "w") as z:
         for l in list:
             z.write(newPath + l,l,zipfile.ZIP_DEFLATED)
             
-for l in list:
-    os.remove(newPath + l)
-        
+    for l in list:
+        os.remove(newPath + l)
+
